+++ cardinal/evm/server_test.go
@@ -81,7 +81,7 @@ func TestServer_SendMessage(t *testing.T) {
 	})
 	assert.NilError(t, w.LoadGameState())
 
-	sender := "0xHelloThere"
+	sender := "0xd5e099c71b797516c10ed0f0d895f429c2781142"
 	personaTag := "foo"
 	// create authorized addresses for the evm transaction's msg sender.
 	ecs.CreatePersonaMsg.AddToQueue(w, ecs.CreatePersona{
