+++ internal/nakama/testsuite_test.go
@@ -7,7 +7,6 @@ import (
 	"io"
 	"math/rand"
 	"net/http"
-	"pkg.world.dev/world-engine/cardinal/ecs"
 	"strings"
 	"testing"
 	"time"
@@ -289,30 +288,6 @@ func TestPersonaTagFieldCannotBeEmpty(t *testing.T) {
 	})
 	assert.NilError(t, err)
 	assert.Equal(t, 400, resp.StatusCode, copyBody(resp))
-}
-
-func TestIsAlphanumericWithUnderscore(t *testing.T) {
-	tests := []struct {
-		input    string
-		expected bool
-	}{
-		{"abc123", true},
-		{"ABC_123", true},
-		{"123", true},
-		{"abc 123", false}, // contains a space
-		{"abc123 ", false}, // contains a trailing space
-		{"abc@123", false}, // contains a special character
-		{"", false},        // empty string
-	}
-
-	for _, test := range tests {
-		t.Run(test.input, func(t *testing.T) {
-			result := ecs.IsAlphanumericWithUnderscore(test.input)
-			if result != test.expected {
-				t.Errorf("Expected %v, but got %v", test.expected, result)
-			}
-		})
-	}
 }
 
 // waitForAcceptedPersonaTag periodically queries the show-persona endpoint until a previously claimed persona tag
